refactor(types): consolidate generic types and utils
Many of our generic or basic types and utils were duplicated across the projects. This commit brings them into `@votingworks/types`.
