+++ apps/bmd/src/utils/machineConfig.ts
@@ -1,5 +1,5 @@
+import { Provider } from '@votingworks/types'
 import {
-  Provider,
   MachineConfig,
   MachineConfigResponse,
   getAppMode,
