+++ apps/election-manager/src/config/types.ts
@@ -2,15 +2,11 @@ import {
   BallotStyle,
   Candidate,
   Contest,
+  Dictionary,
   Election,
+  Optional,
   Precinct,
 } from '@votingworks/types'
-
-// Generic
-export declare type Optional<T> = T | undefined
-export interface Dictionary<T> {
-  [key: string]: Optional<T>
-}
 
 // Events
 export type InputEventFunction = (
